Renaming all of the Domain\Transferlock\Set namespaces to Domain\Set\Transferlock
