+++ lib/event/domain/set/transferlock/fail.php
@@ -16,14 +16,10 @@
  * if not, see https://www.gnu.org/licenses.
  */
 
-namespace Automattic\Domain_Services\Event\Domain\Transferlock\Set;
+namespace Automattic\Domain_Services\Event\Domain\Set\Transferlock;
 
 use Automattic\Domain_Services\{Event};
 
-class Success implements Event\Event_Interface {
-	use Event\Data_Trait, Event\Object_Type_Domain_Trait;
-
-	public function is_locked(): bool {
-		return $this->get_data_by_key( 'event_data.transferlock' ) ?? false;
-	}
+class Fail implements Event\Event_Interface {
+	use Event\Data_Trait, Event\Object_Type_Domain_Trait, Event\Error_Trait;
 }
