+++ test/command/domain-set-transferlock-test.php
@@ -20,12 +20,12 @@ namespace Automattic\Domain_Services\Test\Command;
 
 use Automattic\Domain_Services\{Command, Entity, Test};
 
-class Domain_Transferlock_Set_Test extends Test\Lib\Domain_Services_Client_Test_Case {
+class Domain_Set_Transferlock_Test extends Test\Lib\Domain_Services_Client_Test_Case {
 	use Command\Array_Key_Domain_Trait, Command\Array_Key_Transferlock_Trait;
 
 	public function test_command_instance_success(): void {
 		$mock_command_data = [
-			Command\Command_Interface::COMMAND => 'Domain_Transferlock_Set',
+			Command\Command_Interface::COMMAND => 'Domain_Set_Transferlock',
 			Command\Command_Interface::PARAMS => [
 				self::get_domain_name_array_key() => 'test-domain-name.com',
 				self::get_transferlock_array_key() => true,
@@ -34,10 +34,10 @@ class Domain_Transferlock_Set_Test extends Test\Lib\Domain_Services_Client_Test_
 		];
 
 		$domain = new Entity\Domain_Name( $mock_command_data[ Command\Command_Interface::PARAMS ][ self::get_domain_name_array_key() ] );
-		$command = new Command\Domain\Transferlock\Set( $domain, $mock_command_data[ Command\Command_Interface::PARAMS ][ self::get_transferlock_array_key() ] );
+		$command = new Command\Domain\Set\Transferlock( $domain, $mock_command_data[ Command\Command_Interface::PARAMS ][ self::get_transferlock_array_key() ] );
 		$command->set_client_txn_id( $mock_command_data[ Command\Command_Interface::CLIENT_TXN_ID ] );
 
-		$this->assertInstanceOf( Command\Domain\Transferlock\Set::class, $command );
+		$this->assertInstanceOf( Command\Domain\Set\Transferlock::class, $command );
 
 		$actual_command_array = $command->jsonSerialize();
